1. Rolled in rictic's changes to look up the nodes in a hashmap, instead of iterating through the entire list.
2. Keep track of living nodes as they come alive and die, 
instead of filtering through the entire collection of nodes
everytime.

3. Laid the groundwork to allow new physics engines
to get rid of the n^2 algorithms in favor of n*(n+1)/2

4. Laid the groundwork to remove all the static from 
the code_swarm class, and allow us to pass the class to
the physics engines.

Tested this against all current engines with the sample set,
and everything appears to work as normal. On the next check-in
I hope to add some new physics engines to take advantage
of the speed up.


git-svn-id: http://codeswarm.googlecode.com/svn/trunk@276 eda9c206-d64f-0410-a2f2-67f3fa0499ed